requirements.txt
pyyaml
requests
mlflow
dvc
kubernetes
pytest
scikit-learn
